- enum class Resource extracted - number of enum entries replaced correctly (without Count instance)
