+++ app/src/main/java/de/moyapro/idle/domain/Resource.kt
@@ -1,0 +1,8 @@
+package de.moyapro.idle.domain
+
+enum class Resource(val displayName: String) {
+    EvolutionPoints("evolution points"),
+    Energy("energy"),
+    Water("water"),
+    Minerals("minerals"),
+}
